% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolina_thesis_functions.R
\name{cbb}
\alias{cbb}
\alias{loglikcbb.gamma}
\alias{loglikcbb.mu}
\title{Likelihood function for the chance-corrected beta-binomial model}
\usage{
cbb(a, b, x, k, p0)

loglikcbb.gamma(mu, gamma, y, k, p0)

loglikcbb.mu(gamma, mu, y, k, p0)
}
\arguments{
\item{a}{parameter}

\item{b}{parameter}

\item{x}{number of correct responses}

\item{k}{number of replications}

\item{p0}{guessing probability}

\item{mu}{mu parameter}

\item{gamma}{gamma parameter}

\item{y}{vector of correct answers}
}
\description{
Likelihood function for the chance-corrected beta-binomial model
}
