% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolina_thesis_functions.R
\name{loglikcbb}
\alias{loglikcbb}
\title{Likelihood function for the chance-corrected beta-binomial model}
\usage{
loglikcbb(parameters, y, k, p0)
}
\arguments{
\item{parameters}{mu and gamma parameter vector}

\item{y}{vector of correct answers}

\item{k}{number of replications}

\item{p0}{guessing probability}
}
\description{
Likelihood function for the chance-corrected beta-binomial model
}
