\name{myF}
\alias{myF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A small test function to add stuff}
\description{
  lots of words really
}
\usage{

myF(x)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{this should be some numeric vector of a kind}
}
\details{
  not much to say
  \code{\link{pnorm}}.
}
\value{
 a numeric vector, the input of function + 5
}
%% \references{}
\author{Lucky Luke}
%%\note{ ~~further notes~~
%%
%% ~Make other sections like Warning with \section{Warning }{....} ~
%%}
%%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{

a = 1:10
myF(a)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
