% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolina_thesis_functions.R
\name{optimcbb}
\alias{optimcbb}
\title{Optimize the chance-corrected beta-binomial likelihood}
\usage{
optimcbb(y, k, p0, init = c(0.5, 0.2), max.try = 2)
}
\arguments{
\item{y}{vector of correct answers}

\item{k}{number of replications}

\item{p0}{guessing probability}

\item{init}{vector of initial parameters}

\item{max.try}{maximum number of re-tries}
}
\description{
Optimize the chance-corrected beta-binomial likelihood
}
