% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolina_thesis_functions.R
\name{plotcbb}
\alias{plotcbb}
\title{Plot likelihood confidence region for mu and gamma}
\usage{
plotcbb(y, k, p0, level = 0.95, length.out = 101, p = 2, init = c(0.5,
  0.2), max.try = 2)
}
\arguments{
\item{y}{vector of correct answers}

\item{k}{number of replications}

\item{p0}{guessing probability}

\item{level}{confidence level}

\item{length.out}{number of grid points at each axis}

\item{p}{df for the chi-square approximation (between 1 and 2)}

\item{init}{vector of initial parameters}

\item{max.try}{maximum number of re-tries}
}
\description{
Plot likelihood confidence region for mu and gamma
}
