% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolina_thesis_functions.R
\name{profilecbb}
\alias{profilecbb}
\title{Profile CBB likelihood function}
\usage{
profilecbb(y, k, p0, plot.profile = TRUE, level = 0.95, length.out = 101,
  init = c(0.5, 0.2), max.try = 2)
}
\arguments{
\item{y}{vector of correct answers}

\item{k}{number of replications}

\item{p0}{guessing probability}

\item{plot.profile}{logical; plot the profile likelihoods?}

\item{level}{confidence level}

\item{length.out}{number of grid points at each axis}

\item{init}{vector of initial parameters}

\item{max.try}{maximum number of re-tries}
}
\description{
Profile CBB likelihood function
}
