% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubinRules.R
\name{rubinRules}
\alias{rubinRules}
\title{Assessing Rubin's rules in design of observational studies}
\usage{
rubinRules(data, Treatment, matchscore = "ps", covlist)
}
\arguments{
\item{data}{a dataframe which must have either a propensity score(PS) column or a transformed version of the PS}

\item{Treatment}{variable contianing treatment assignments with 1 representing treatment and 0 represinting no treatment}

\item{matchscore}{Name of the score used to match.can take the values "ps" or "linps"}

\item{covlist}{a vector containing a list of variables in your study}
}
\description{
This function generates a plot to assess rubin's rules. The ratio of residual and variances, Rule 1 and 2 respectively, are displayed at the top of the plot.
}
\examples{
 covlist1=c("covA", "covB", "covC", "covD", "covE", "covF.Middle", "covF.High", "Asqr","BC", "BD") # creating a vector of variables
# calling function
rubinRules(data=toy.matchedsample,Treatment="treated",covlist=covlist1)
}
\keyword{Rubin}
\keyword{rules}

