% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensbinary.R
\name{sensBinary}
\alias{sensBinary}
\title{Sensitivity Analysis for A Simple Comparison for binary outcomes
This function allows you to assess how sensitive your results are to unmeasured variable.}
\usage{
sensBinary(data, StartGamma, EndGamma, Gammainterval, outcome, exp)
}
\arguments{
\item{data}{A matched sample}

\item{Gammainterval}{interval between two hidden bias to be assessed}

\item{outcome}{The outcome variable}

\item{exp}{A variables defining exposure group}

\item{Gamma}{Bias to be assessed}

\item{alpha}{Significance level}
}
\description{
Sensitivity Analysis for A Simple Comparison for binary outcomes
This function allows you to assess how sensitive your results are to unmeasured variable.
}
\examples{
 data("Telemedicine.matchedsample") # Reading in the data produced by a matching method

sensBinary(data=Telemedicine.matchedsample,StartGamma=1,EndGamma=1.9,Gammainterval=0.01,outcome ="Readmission", exp="Telehealth")
}
\references{
Section 4.4.8. of Rosenbaum PR (2002) Observational Studies, 2nd Edition.
}
\keyword{Sensitivity}

