% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensCont.R
\name{sensCont}
\alias{sensCont}
\title{Sensitivity Analysis for A Simple Comparison for continious data}
\usage{
sensCont(data, exp, outcome, Gamma = 6, Gammainterval = 0.01,
  alpha = 0.05)
}
\arguments{
\item{data}{A matched sample}

\item{exp}{A variables defining exposure group}

\item{outcome}{The outcome variable}

\item{Gamma}{Bias to be assessed}

\item{Gammainterval}{interval between two hidden bias to be assessed}

\item{alpha}{Significance level}

\item{failtime}{Time to event}
}
\description{
This function allows you to assess how sensitive your results are to unmeasured variable.
}
\examples{
 data("Telemedicine.matchedsample") # Reading in the data produced by a matching method

sensCont(data=Telemedicine.matchedsample,exp="Telehealth.n",outcome = "GFRoneyr",Gamma=1.5,Gammainterval = 0.06)
}
\references{
Section 4.4.8. of Rosenbaum PR (2002) Observational Studies, 2nd Edition.
}
\keyword{Sensitivity}

