\name{sensR-deprecated}
\alias{sensR-deprecated}
%% \alias{clls}

\title{Deprecated Functions in sensR Package}

\description{
  These functions are provided for compatibility with older versions of
  the \pkg{sensR} package only, and may be removed eventually.
}

\details{
  The \code{clls} function is deprecated and users encouraged to use
  \code{\link[ordinal]{clm}} instead from the \pkg{ordinal} package.
}
